Work in progress: adding support for websockets (JSR-356).
Almost done. The only problem now is that it doesn't work in Glassfish 4. For some esoteric reason, the user properties stored via ServerEndpointConfig.getUserProperties() seem to fall into a black hole never to return. They should be present in Session.getUserProperties() when Dagger.onOpen() is called, but guess what, they're not. Curious how it works fine in Tomcat 8.